dom/performance/tests/test_worker_performance_entries.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>PerformanceResouceTiming in workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
var worker = new Worker('test_worker_performance_entries.js');
worker.onmessage = function(event) {
  if (event.data.type == "check") {
    ok(event.data.status, event.data.msg);
    return;
  }

  if (event.data.type == "finish") {
    SimpleTest.finish();
    return;
  }

  ok(false, "?!?");
}

</script>
</body>
</html>
